\name{timeperiodsR-package}
\alias{timeperiodsR-package}
\alias{timeperiodsR}
\docType{package}
\title{
\packageTitle{timeperiodsR}
}
\description{
\packageDescription{timeperiodsR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{timeperiodsR}
\packageIndices{timeperiodsR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{timeperiodsR}

Maintainer: \packageMaintainer{timeperiodsR}
}
