\name{timeperiodsR-package}
\alias{timeperiodsR-package}
\alias{timeperiodsR}
\docType{package}
\title{
\packageTitle{timeperiodsR}
}
\description{
\packageDescription{timeperiodsR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{timeperiodsR}
\packageIndices{timeperiodsR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{timeperiodsR}

Maintainer: \packageMaintainer{timeperiodsR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
